Update AutoClip.py
I handled the exceptions as best as I could, as there was really nothing more I could do at this point in time. The script functions exactly as it purports to, the only errors are from rapidly deleting tracks due to a race condition between removing event listeners from the track itself and removing the event listener of a track from the song() properties. I would be really grateful to anyone in the community who could help out with this! you'd be making the script more robust in your contribution.