Update AutoClip.py
Fixed the error message by properly navigating the output of a race condition, where the none.none(ClipSlot) type properly gets converted to the none pointer in C++ and so we can effectively handle this intermediary step without that blasted red error message of doom. but actually low key that terminal slaps and got me inspired to make remote scripts in the first place.